Fix for issue #339: undefined behavior in impl_sse/p7_oprofile.c
biased_byteify() converts a log-odds score as a float to a cost as a
uint8_t. It scales and rounds the score to an exact integer value (in
a float), adds a bias, and casts to uint8_t. Because the bias is
previously determined as the minimum scaled/rounded score, then after
the bias is added, we know sc has a nonnegative exact integral value
(0., 1., 2., ...) that is safe to cast to uint8_t (after a check for
saturation > 255).

However, the code was casting sc to uint8_t *before* the bias was
added, while sc could still contain a negative value. (I think the
reason was that the bias itself is already stored as uint8_t.)
Casting a negative float value to an unsigned integer type results in
undefined behavior according to the C standard. If the compiler casts
the float to signed int first, then uses twos-complement to convert
signed to unsigned, the code works. But if the compiler truncates
negative float values to zero first, and then casts to unsigned, the
code fails.

Our NEON vector code (by Martin Larradale) and VMX vector code (by
Michael Farrar) already had workarounds implemented. Only the SSE
vector code (by me) had the problem.

The fix adds the bias (as a float) while sc is still a float,
truncates it to <= 255, then casts to uint8_t.

I copied this corrected implementation to the NEON and VMX
implementations, removing the previous workarounds.

[xref SRE:0721-h3-iss339-byteify]